+++ app/utils/idNumber.ts
@@ -1,0 +1,64 @@
+export const validateIdNumber = (personType: string, idNumber: string): boolean => {
+  if (personType === "fisica") {
+    return validateCPF(idNumber);
+  } else if (personType === "juridica") {
+    return validateCNPJ(idNumber);
+  } else {
+    return false;
+  }
+};
+
+// Validação de CPF
+export const validateCPF = (cpf: string): boolean => {
+  cpf = cpf.replace(/[^\d]+/g, "");
+  if (cpf.length !== 11 || /^(\d)\1+$/.test(cpf)) return false;
+
+  let sum = 0;
+  for (let i = 0; i < 9; i++) sum += parseInt(cpf.charAt(i)) * (10 - i);
+  let rev = 11 - (sum % 11);
+  if (rev === 10 || rev === 11) rev = 0;
+  if (rev !== parseInt(cpf.charAt(9))) return false;
+
+  sum = 0;
+  for (let i = 0; i < 10; i++) sum += parseInt(cpf.charAt(i)) * (11 - i);
+  rev = 11 - (sum % 11);
+  if (rev === 10 || rev === 11) rev = 0;
+  if (rev !== parseInt(cpf.charAt(10))) return false;
+
+  return true;
+};
+
+// Validação de CNPJ
+export const validateCNPJ = (cnpj: string): boolean => {
+  cnpj = cnpj.replace(/[^\d]+/g, "");
+  if (cnpj.length !== 14) return false;
+
+  // Elimina CNPJs invalidos conhecidos
+  if (/^(.)\1+$/.test(cnpj)) return false;
+
+  // Valida DVs
+  let length = cnpj.length - 2;
+  let numbers = cnpj.substring(0, length);
+  const digits = cnpj.substring(length);
+  let sum = 0;
+  let pos = length - 7;
+  for (let i = length; i >= 1; i--) {
+    sum += parseInt(numbers.charAt(length - i)) * pos--;
+    if (pos < 2) pos = 9;
+  }
+  let result = sum % 11 < 2 ? 0 : 11 - (sum % 11);
+  if (result !== parseInt(digits.charAt(0))) return false;
+
+  length += 1;
+  numbers = cnpj.substring(0, length);
+  sum = 0;
+  let pos2 = length - 7;
+  for (let i = length; i >= 1; i--) {
+    sum += parseInt(numbers.charAt(length - i)) * pos2--;
+    if (pos2 < 2) pos2 = 9;
+  }
+  result = sum % 11 < 2 ? 0 : 11 - (sum % 11);
+  if (result !== parseInt(digits.charAt(1))) return false;
+
+  return true;
+};
